Use waveguide plugin from Tidy3D mode solver
Almost all features from the previous solver are available in some form.
We're missing the implementation of the index dictionary that was
overlaid in the previous version to enable the user to set a
non homogeneous index profile over the geometry. That can be enabled
through `CustomMedium` in the future.

There are a few improvements from the previous version, most notably the
use of a non-uniform grid.

This commit solves issues from both
https://github.com/gdsfactory/gdsfactory/issues/1056 and
https://github.com/flexcompute/tidy3d/issues/652.

Signed-off-by: Lucas Heitzmann Gabrielli <lucas@flexcompute.com>