Removal of unnecessary empty gf.Component() while using gf.c.text(...).
Each line of the text was creating an unnamed gf.Component(), which is lost in the void once the final text Component is flattened. (Not even sure the flatten is still necessary). 
For texts with a lot of lines, this was generating a lot of empty gf.Component(), floading the cache.